Change badges id to UUID and create a common app Schema
This allows to keep a common schema across all module schemas. Read more
about
[here](https://hexdocs.pm/ecto/Ecto.Schema.html#module-schema-attributes).

This also allows to not worry about the path of an image based on the ID
because UUID are generating before inserting the data to the db. Read
[this](https://hexdocs.pm/waffle_ecto/filepath-with-id.html) to
understand why badge image does not require creation to be a transaction
anymore.